requirements.txt
google-cloud>=0.34.0
protobuf>=3.18.1
kafka-python>=2.0
colorama==0.4.3
